// generated/addon/angular/src/app/api-models/cieb-trans-accum-orig-sys-limt.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class CiebTransAccumOrigSysLimt {
  seqSysId: number;
  seqTransId: number;
  sysLimitName: string;
  sysXLimVal: string;
  sysAccumPeriodTy: string;
  sysCarryoverType: string;
  sysCarryoverValue: number;
  cntrDollarInd: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}